api/pyproject.toml
[project]
name = "hyper-trader-api"
version = "1.0.0"
description = "HyperTrader SaaS API - Multi-tenant trading bot management platform"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Andrei Nasui" }
]

dependencies = [
    # Web Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    # Configuration & Validation
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    # YAML parsing
    "pyyaml>=6.0.0",
    # Security & Auth
    "bcrypt>=4.0.0",
    "cryptography>=41.0.0",
    "python-jose[cryptography]>=3.3.0",
    # HTTP Client
    "httpx>=0.25.0",
    # File uploads
    "python-multipart>=0.0.6",
    # Kubernetes
    "kubernetes>=28.0.0",
    # Templating
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
api = "uvicorn:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["api"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
